Update seq2seq + attention (#347)
* more

* more
