+++ chapter_recurrent-neural-networks/machine-translation.md
@@ -86,7 +86,7 @@ len(src_vocab)
 
 In language models, each example is a `num_steps` length sequence from the corpus, which may be a segment of a sentence, or span over multiple sentences. In machine translation, an example should contain a pair of source sentence and target sentence. These sentences might have different lengths, while we need same length examples to form a mini-batch. 
 
-One way to solve this problem is that we if a sentence is longer than `num_steps`, we trim it's length, otherwise pad with a special &lt;pad&gt; token to meet the length. Therefore we could transform any sentence to a fixed length. 
+One way to solve this problem is that we if a sentence is longer than `num_steps`, we trim it's length, otherwise pad with a special &lt;pad&gt; token to meet the length. Therefore we could transform any sentence to a fixed length.
 
 ```{.python .input  n=11}
 # Save to the d2l package.
@@ -114,13 +114,13 @@ Then we can construct mini-batches based on these arrays.
 
 ## Put All Things Together
 
-Finally, we define the function `load_data_nmt` to return the data iterator with the vocabularies for source language and target language. 
+Finally, we define the function `load_data_nmt` to return the data iterator with the vocabularies for source language and target language.
 
 ```{.python .input  n=13}
 # Save to the d2l package.
 def load_data_nmt(batch_size, num_steps, num_examples=1000):
     text = preprocess_nmt(read_data_nmt())
-    source, target = tokenize_nmt(text)
+    source, target = tokenize_nmt(text, num_examples)
     src_vocab = d2l.Vocab(source, min_freq=3, use_special_tokens=True)
     tgt_vocab = d2l.Vocab(target, min_freq=3, use_special_tokens=True)
     src_array, src_valid_len = build_array(
