+++ d2l/__init__.py
@@ -1,7 +1,6 @@
 from .base import *
 from .figure import *
 from .data import *
-from .model import *
 from .train import *
 
 from .d2l import *
