+++ d2l/d2l.py
@@ -14,7 +14,7 @@ import numpy as np
 import math
 from matplotlib import pyplot as plt
 from mxnet import nd, autograd, gluon, init, context, image
-from mxnet.gluon import nn
+from mxnet.gluon import nn, rnn
 import random
 import re
 import time
@@ -631,7 +631,7 @@ def build_array(lines, vocab, num_steps, is_source):
 # Defined in file: ./chapter_recurrent-neural-networks/machine-translation.md
 def load_data_nmt(batch_size, num_steps, num_examples=1000):
     text = preprocess_nmt(read_data_nmt())
-    source, target = tokenize_nmt(text)
+    source, target = tokenize_nmt(text, num_examples)
     src_vocab = d2l.Vocab(source, min_freq=3, use_special_tokens=True)
     tgt_vocab = d2l.Vocab(target, min_freq=3, use_special_tokens=True)
     src_array, src_valid_len = build_array(
@@ -641,6 +641,190 @@ def load_data_nmt(batch_size, num_steps, num_examples=1000):
     data_arrays = (src_array, src_valid_len, tgt_array, tgt_valid_len)
     data_iter = d2l.load_array(data_arrays, batch_size)
     return src_vocab, tgt_vocab, data_iter
+
+# Defined in file: ./chapter_recurrent-neural-networks/encoder-decoder.md
+class Encoder(nn.Block):
+    """The base encoder interface for the encoder-decoder architecture."""
+    def __init__(self, **kwargs):
+        super(Encoder, self).__init__(**kwargs)
+
+    def forward(self, X):
+        raise NotImplementedError
+
+# Defined in file: ./chapter_recurrent-neural-networks/encoder-decoder.md
+class Decoder(nn.Block):
+    """The base decoder interface for the encoder-decoder archtecture."""
+    def __init__(self, **kwargs):
+        super(Decoder, self).__init__(**kwargs)
+
+    def init_state(self, enc_outputs, *args):
+        raise NotImplementedError
+
+    def forward(self, X, state):
+        raise NotImplementedError
+
+# Defined in file: ./chapter_recurrent-neural-networks/encoder-decoder.md
+class EncoderDecoder(nn.Block):
+    """The base class for the encoder-decoder architecture."""
+    def __init__(self, encoder, decoder, **kwargs):
+        super(EncoderDecoder, self).__init__(**kwargs)
+        self.encoder = encoder
+        self.decoder = decoder
+
+    def forward(self, enc_X, dec_X, *args):
+        enc_outputs = self.encoder(enc_X, *args)
+        dec_state = self.decoder.init_state(enc_outputs, *args)
+        return self.decoder(dec_X, dec_state)
+
+# Defined in file: ./chapter_recurrent-neural-networks/seq2seq.md
+class Seq2SeqEncoder(d2l.Encoder):
+    def __init__(self, vocab_size, embed_size, num_hiddens, num_layers,
+                 dropout=0, **kwargs):
+        super(Seq2SeqEncoder, self).__init__(**kwargs)
+        self.embedding = nn.Embedding(vocab_size, embed_size)
+        self.rnn = rnn.LSTM(num_hiddens, num_layers, dropout=dropout)
+
+    def forward(self, X, *args):
+        X = self.embedding(X) # X shape: (batch_size, seq_len, embed_size)
+        X = X.swapaxes(0, 1)  # RNN needs first axes to be time
+        state = self.rnn.begin_state(batch_size=X.shape[1], ctx=X.context)
+        out, state = self.rnn(X, state)
+        # The shape of out is (seq_len, batch_size, num_hiddens).
+        # state contains the hidden state and the memory cell
+        # of the last time step, the shape is (num_layers, batch_size, num_hiddens)
+        return out, state
+
+# Defined in file: ./chapter_recurrent-neural-networks/seq2seq.md
+class Seq2SeqDecoder(d2l.Decoder):
+    def __init__(self, vocab_size, embed_size, num_hiddens, num_layers,
+                 dropout=0, **kwargs):
+        super(Seq2SeqDecoder, self).__init__(**kwargs)
+        self.embedding = nn.Embedding(vocab_size, embed_size)
+        self.rnn = rnn.LSTM(num_hiddens, num_layers, dropout=dropout)
+        self.dense = nn.Dense(vocab_size, flatten=False)
+
+    def init_state(self, enc_outputs, *args):
+        return enc_outputs[1]
+
+    def forward(self, X, state):
+        X = self.embedding(X).swapaxes(0, 1)
+        out, state = self.rnn(X, state)
+        # Make the batch to be the first dimension to simplify loss computation.
+        out = self.dense(out).swapaxes(0, 1)
+        return out, state
+
+# Defined in file: ./chapter_recurrent-neural-networks/seq2seq.md
+class MaskedSoftmaxCELoss(gluon.loss.SoftmaxCELoss):
+    # pred shape: (batch_size, seq_len, vocab_size)
+    # label shape: (batch_size, seq_len)
+    # valid_length shape: (batch_size, )
+    def forward(self, pred, label, valid_length):
+        # the sample weights shape should be (batch_size, seq_len, 1)
+        weights = nd.ones_like(label).expand_dims(axis=-1)
+        weights = nd.SequenceMask(weights, valid_length, True, axis=1)
+        return super(MaskedSoftmaxCELoss, self).forward(pred, label, weights)
+
+# Defined in file: ./chapter_recurrent-neural-networks/seq2seq.md
+def train_s2s_ch8(model, data_iter, lr, num_epochs, ctx):
+    model.initialize(init.Xavier(), force_reinit=True, ctx=ctx)
+    trainer = gluon.Trainer(model.collect_params(),
+                            'adam', {'learning_rate': lr})
+    loss = MaskedSoftmaxCELoss()
+    #tic = time.time()
+    animator = d2l.Animator(xlabel='epoch', ylabel='loss', 
+                            xlim=[1, num_epochs], ylim=[0, 0.25])
+    for epoch in range(1, num_epochs+1):
+        timer, metric = d2l.Timer(), d2l.Accumulator(2)  # loss_sum, num_tokens
+        for batch in data_iter:
+            X, X_vlen, Y, Y_vlen = [x.as_in_context(ctx) for x in batch]
+            Y_input, Y_label, Y_vlen = Y[:,:-1], Y[:,1:], Y_vlen-1
+            with autograd.record():
+                Y_hat, _ = model(X, Y_input, X_vlen, Y_vlen)
+                l = loss(Y_hat, Y_label, Y_vlen)
+            l.backward()
+            d2l.grad_clipping(model, 1)
+            num_tokens = Y_vlen.sum().asscalar()
+            trainer.step(num_tokens)
+            metric.add((l.sum().asscalar(), num_tokens))
+        if epoch % 10 == 0:
+            animator.add(epoch, metric[0]/metric[1])
+    print('loss %.3f, %d tokens/sec on %s ' % (
+        metric[0]/metric[1], metric[1]/timer.stop(), ctx))
+
+# Defined in file: ./chapter_recurrent-neural-networks/seq2seq.md
+def predict_s2s_ch8(model, src_sentence, src_vocab, tgt_vocab, num_steps, ctx):
+    src_tokens = src_vocab[src_sentence.lower().split(' ')]
+    enc_valid_length = nd.array([len(src_tokens)], ctx=ctx)
+    src_tokens = d2l.trim_pad(src_tokens, num_steps, src_vocab.pad)
+    enc_X = nd.array(src_tokens, ctx=ctx)
+    # add the batch_size dimension.
+    enc_outputs = model.encoder(enc_X.expand_dims(axis=0), enc_valid_length)
+    dec_state = model.decoder.init_state(enc_outputs, enc_valid_length)
+    dec_X = nd.array([tgt_vocab.bos], ctx=ctx).expand_dims(axis=0)
+    predict_tokens = []
+    for _ in range(num_steps):
+        Y, dec_state = model.decoder(dec_X, dec_state)
+        # The token with highest score is used as the next time step input.
+        dec_X = Y.argmax(axis=2)
+        py = dec_X.squeeze(axis=0).astype('int32').asscalar()
+        if py == tgt_vocab.eos:
+            break
+        predict_tokens.append(py)
+    return ' '.join(tgt_vocab.to_tokens(predict_tokens))
+
+# Defined in file: ./chapter_attention-mechanism/attention.md
+def masked_softmax(X, valid_length):
+    # X: 3-D tensor, valid_length: 1-D or 2-D tensor
+    if valid_length is None:
+        return X.softmax()
+    else:
+        shape = X.shape
+        if valid_length.ndim == 1:
+            valid_length = valid_length.repeat(shape[1], axis=0)
+        else:
+            valid_length = valid_length.reshape((-1,))
+        # fill masked elements with a large negative, whose exp is 0
+        X = nd.SequenceMask(X.reshape((-1, shape[-1])), valid_length, True,
+                            axis=1, value=-1e6)
+        return X.softmax().reshape(shape)
+
+# Defined in file: ./chapter_attention-mechanism/attention.md
+class DotProductAttention(nn.Block): 
+    def __init__(self, dropout, **kwargs):
+        super(DotProductAttention, self).__init__(**kwargs)
+        self.dropout = nn.Dropout(dropout)
+
+    # query: (batch_size, #queries, d)
+    # key: (batch_size, #kv_pairs, d)
+    # value: (batch_size, #kv_pairs, dim_v)
+    # valid_length: either (batch_size, ) or (batch_size, xx)
+    def forward(self, query, key, value, valid_length=None):
+        d = query.shape[-1]
+        # set transpose_b=True to swap the last two dimensions of key
+        scores = nd.batch_dot(query, key, transpose_b=True) / math.sqrt(d)
+        attention_weights = self.dropout(masked_softmax(scores, valid_length))
+        return nd.batch_dot(attention_weights, value)
+
+# Defined in file: ./chapter_attention-mechanism/attention.md
+class MLPAttention(nn.Block):  
+    def __init__(self, units, dropout, **kwargs):
+        super(MLPAttention, self).__init__(**kwargs)
+        # Use flatten=True to keep query's and key's 3-D shapes.
+        self.W_k = nn.Dense(units, activation='tanh',
+                            use_bias=False, flatten=False)
+        self.W_q = nn.Dense(units, activation='tanh',
+                            use_bias=False, flatten=False)
+        self.v = nn.Dense(1, use_bias=False, flatten=False)
+        self.dropout = nn.Dropout(dropout)
+
+    def forward(self, query, key, value, valid_length):
+        query, key = self.W_k(query), self.W_q(key)
+        # expand query to (batch_size, #querys, 1, units), and key to
+        # (batch_size, 1, #kv_pairs, units). Then plus them with broadcast.
+        features = query.expand_dims(axis=2) + key.expand_dims(axis=1)
+        scores = self.v(features).squeeze(axis=-1)
+        attention_weights = self.dropout(masked_softmax(scores, valid_length))
+        return nd.batch_dot(attention_weights, value)
 
 # Defined in file: ./chapter_optimization/optimization-intro.md
 def annotate(text, xy, xytext):
