--- d2l/model.py
@@ -1,120 +1,0 @@
-"""The model module contains neural network building blocks"""
-import math
-from mxnet import nd
-from mxnet.gluon import nn, rnn, loss as gloss
-
-
-
-
-class RNNModel(nn.Block):
-    """RNN model."""
-
-    def __init__(self, rnn_layer, vocab_size, **kwargs):
-        super(RNNModel, self).__init__(**kwargs)
-        self.rnn = rnn_layer
-        self.vocab_size = vocab_size
-        self.dense = nn.Dense(vocab_size)
-
-    def forward(self, inputs, state):
-        """Forward function"""
-        X = nd.one_hot(inputs.T, self.vocab_size)
-        Y, state = self.rnn(X, state)
-        output = self.dense(Y.reshape((-1, Y.shape[-1])))
-        return output, state
-
-    def begin_state(self, *args, **kwargs):
-        """Return the begin state"""
-        return self.rnn.begin_state(*args, **kwargs)
-
-class Encoder(nn.Block):
-    """The base encoder interface for the encoder-decoder architecture."""
-    def __init__(self, **kwargs):
-        super(Encoder, self).__init__(**kwargs)
-
-    def forward(self, X, *args):
-        """Forward function"""
-        raise NotImplementedError
-
-class Decoder(nn.Block):
-    """The base decoder interface for the encoder-decoder archtecture."""
-    def __init__(self, **kwargs):
-        super(Decoder, self).__init__(**kwargs)
-
-    def init_state(self, enc_outputs, *args):
-        """Return the begin state"""
-        raise NotImplementedError
-
-    def forward(self, X, state):
-        """Forward function"""
-        raise NotImplementedError
-
-class EncoderDecoder(nn.Block):
-    """The base class for the encoder-decoder architecture."""
-    def __init__(self, encoder, decoder, **kwargs):
-        super(EncoderDecoder, self).__init__(**kwargs)
-        self.encoder = encoder
-        self.decoder = decoder
-
-    def forward(self, enc_X, dec_X, *args):
-        """Forward function"""
-        enc_outputs = self.encoder(enc_X, *args)
-        dec_state = self.decoder.init_state(enc_outputs, *args)
-        return self.decoder(dec_X, dec_state)
-
-class Seq2SeqEncoder(Encoder):
-    def __init__(self, vocab_size, embed_size, num_hiddens, num_layers,
-                 dropout=0, **kwargs):
-        super(Seq2SeqEncoder, self).__init__(**kwargs)
-        self.embedding = nn.Embedding(vocab_size, embed_size)
-        self.rnn = rnn.LSTM(num_hiddens, num_layers, dropout=dropout)
-
-    def forward(self, X, *args):
-        """Forward function"""
-        X = self.embedding(X)
-        X = X.swapaxes(0, 1)
-        state = self.rnn.begin_state(batch_size=X.shape[1], ctx=X.context)
-        out, state = self.rnn(X, state)
-        return out, state
-
-def masked_softmax(X, valid_length):
-    if valid_length is None:
-        return X.softmax()
-    else:
-        shape = X.shape
-        if valid_length.ndim == 1:
-            valid_length = valid_length.repeat(shape[1], axis=0)
-        else:
-            valid_length = valid_length.reshape((-1,))
-        X = nd.SequenceMask(X.reshape((-1, shape[-1])), valid_length, True,
-                            axis=1, value=-1e6)
-        return X.softmax().reshape(shape)
-
-class DotProductAttention(nn.Block):
-    def __init__(self, dropout, **kwargs):
-        super(DotProductAttention, self).__init__(**kwargs)
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, query, key, value, valid_length=None):
-        """Forward function"""
-        d = query.shape[-1]
-        scores = nd.batch_dot(query, key, transpose_b=True) / math.sqrt(d)
-        attention_weights = self.dropout(masked_softmax(scores, valid_length))
-        return nd.batch_dot(attention_weights, value)
-
-class MLPAttention(nn.Block):
-    def __init__(self, units, dropout, **kwargs):
-        super(MLPAttention, self).__init__(**kwargs)
-        self.W_k = nn.Dense(units, activation='tanh',
-                            use_bias=False, flatten=False)
-        self.W_q = nn.Dense(units, activation='tanh',
-                            use_bias=False, flatten=False)
-        self.v = nn.Dense(1, use_bias=False, flatten=False)
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, query, key, value, valid_length):
-        """Forward function"""
-        query, key = self.W_k(query), self.W_q(key)
-        features = query.expand_dims(axis=2) + key.expand_dims(axis=1)
-        scores = self.v(features).squeeze(axis=-1)
-        attention_weights = self.dropout(masked_softmax(scores, valid_length))
-        return nd.batch_dot(attention_weights, value)
